Remove condition namespace, refactor and simplify naming
